+++ CHANGELOG.md
@@ -53,6 +53,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed: set missing require for `phoneNumber` on the Checkout
 
 
+### Changed / Improved
+- Rearranged components of MyAccount page in one folder(#632)
+
 ## [1.0.4] - 04.01.2020
 
 ### Added